refactor(article,series): integrate series webmvc adapter to article webmvc adapter
