+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/SeriesQueryApi.java
@@ -1,4 +1,4 @@
-package nettee.series.driving.web;
+package nettee.blole.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -6,13 +6,13 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import nettee.blole.article.web.dto.SeriesQueryDto.SeriesDetailResponse;
+import nettee.blole.article.web.dto.SeriesQueryDto.SeriesSummaryResponse;
 import nettee.blolet.article.application.usecase.SeriesReadUseCase;
 import nettee.blolet.article.application.usecase.SeriesVisitUseCase;
 import nettee.blolet.article.readmodel.SeriesQueryModels.SeriesDetail;
 import nettee.blolet.jwt.filter.annotation.AuthUser;
 import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
-import nettee.series.driving.web.dto.SeriesQueryDto.SeriesDetailResponse;
-import nettee.series.driving.web.dto.SeriesQueryDto.SeriesSummaryResponse;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
