+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/mapper/SeriesArticleDtoMapper.java
@@ -1,7 +1,7 @@
-package nettee.series.article.driving.web.mapper;
+package nettee.blole.article.web.mapper;
 
+import nettee.blole.article.web.dto.SeriesArticleCommandDto.SeriesArticleCreateCommand;
 import nettee.blolet.article.domain.SeriesArticle;
-import nettee.series.article.driving.web.dto.SeriesArticleCommandDto.SeriesArticleCreateCommand;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
